test(builder): refactor JsonDb source specs with shared helpers
